♻️ refactor: EvalLogItem 분리
