+++ app/src/pages/EvalLogSearchPage/EvalLogItem/FlagLabel.tsx
@@ -1,0 +1,19 @@
+import { EvalLogLabel, EvalLogLabelType } from '@components/common/Label';
+import { useTheme } from '@emotion/react';
+import { snakeCaseFormatter } from '@utils/formatters';
+
+type FlagLabelProps = {
+  name: string;
+  isPositive: boolean;
+};
+
+export const FlagLabel = ({ name, isPositive }: FlagLabelProps) => {
+  const theme = useTheme();
+  const type: EvalLogLabelType = isPositive ? 'positive' : 'negative';
+
+  return (
+    <EvalLogLabel type={type} fontWeight={theme.fonts.weight.bold}>
+      {snakeCaseFormatter(name)}
+    </EvalLogLabel>
+  );
+};
